django-original-template-staging/requirements.txt
coverage==5.1
django_uwsgi==0.2.2
django-compressor==2.4
django-debug-toolbar==2.2
django-filter==2.3.0
django-import-export==2.2.0
django-oauth-toolkit==1.3.2
django-rest-swagger==2.2.0
django-sass-processor==0.8
Django==3.0.7
djangorestframework==3.11
drf-url-filters==0.5.1
enum34==1.1.10
fabric==2.5.0
flake8==3.8.2
markdown==3.2.2
mysqlclient==1.4.6
numpy==1.18.5
pyflakes==2.1.1
six==1.15.0
south==1.0.2
sqlparse==0.3.1
uWSGI==2.0.18

##################################################################
#### These should be included on every Django/Python project #####
##################################################################
# security check packages
# https://github.com/pyupio/safety
safety==1.8.7

# https://github.com/bartTC/pip-check/
pip-check==2.6.0

# https://github.com/PyCQA/bandit
bandit==1.6.2

# license check packages
# https://pypi.org/project/pip-licenses/
pip-licenses==2.1.1
